refactor: 패키지 간 순환 참조하지 않도록 패키지 구조 변경
