--- src/main/java/ladder/domain/factory/LineFactory.java
@@ -1,7 +1,0 @@
-package ladder.domain.factory;
-
-@FunctionalInterface
-public interface LineFactory {
-
-    ILine create(int playerCount);
-}
