+++ src/main/java/ladder/domain/line/ILine.java
@@ -1,0 +1,12 @@
+package ladder.domain.line;
+
+import ladder.domain.position.Position;
+
+import java.util.List;
+
+public interface ILine {
+
+    Position nextPosition(Position position);
+
+    List<Boolean> bridges();
+}
