+++ src/main/java/ladder/domain/line/LineFactory.java
@@ -1,0 +1,7 @@
+package ladder.domain.line;
+
+@FunctionalInterface
+public interface LineFactory {
+
+    ILine create(int playerCount);
+}
