+++ src/main/java/ladder/view/InputView.java
@@ -1,6 +1,12 @@
 package ladder.view;
 
-import ladder.domain.*;
+import ladder.domain.ladderresult.LadderResult;
+import ladder.domain.ladderresult.LadderResults;
+import ladder.domain.player.Player;
+import ladder.domain.player.PlayerName;
+import ladder.domain.player.PlayerNames;
+import ladder.domain.player.Players;
+import ladder.domain.position.Position;
 
 import java.util.Arrays;
 import java.util.Scanner;
