+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,6 +1,10 @@
 package ladder.domain;
 
-import ladder.domain.v1.Line;
+import ladder.domain.line.v1.Line;
+import ladder.domain.player.Player;
+import ladder.domain.player.PlayerName;
+import ladder.domain.player.Players;
+import ladder.domain.position.Position;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
